nbg-benchmark: introduce a benchmark and MAC compute tool
Introducing a tool for benchmarking network binary based models
and computing the MAC utilization on the NPU by the model. This
tool is based only the OpenVX API directly with no need to use
NBG-parser or any other intermediate lib. The user has the
choice to either load ACUITY generated .tensor file as an input
or use a random input. If no caseMAC is specified as an args,
the tool does only the inference time benchmarking.

Signed-off-by: Othmane AHL ZOUAOUI <othmane.ahlzouaoui@st.com>