onnxruntime: xnnpack: fix mcpu cortex-a35 build failure
The "-mcpu=" switch is introduced externaly into CMake by Yocto
buildsystem toolchain. This switch is not compatible with XNNPACK
build mechanism and causes the XNNPACK compilation break due to
"unsupported instructions" on cortex-a35 based cpu.
This switch needs to be removed for XNNPACK on order to isolate
the changes only for XNNPACK and its depencencies, a subfolder is
introduced through this patch.

Signed-off-by: Othmane AHL ZOUAOUI <othmane.ahlzouaoui@st.com>