tensorflow-lite : xnnpack : fix mcpu cortex-a35 build failure
The "-mcpu=" switch might be introduced externaly into CMake: either in <LANG>_FLAGS or
as part of CC, CXX, ASM environmental variables (to be stored in CMAKE_<LANG>_COMPILER_ARG1).
This switch is not compatible with XNNPACK build mechanism and causes the XNNPACK compilation
break due to "unsupported instructions". This switch needs to be removed for XNNPACK
In order to isolate the changes only for XNNPACK and its depencencies, a subfolder is
introduced.

Signed-off-by: Maxence GUILHIN <maxence.guilhin@st.com>